chore(edit-content): validate errors #30061
