+++ core-web/libs/edit-content/src/lib/fields/dot-edit-content-file-field/components/dot-form-import-url/dot-form-import-url.component.html
@@ -32,7 +32,7 @@
             type="button"
             aria-label="Cancel button"
             data-testId="cancel-button" />
-        <div>
+        <div [style]="{ width: '6.85rem' }">
             @if (store.isLoading()) {
                 <p-button
                     [icon]="'pi pi-spin pi-spinner'"
